refactor: adds a new metadata base class
