+++ tests/test_metadataJSON.py
@@ -1,12 +1,9 @@
-import json
 import os
 from unittest import TestCase
-from unittest.mock import patch, MagicMock, ANY, Mock
+from unittest.mock import patch, MagicMock, Mock
 from collections import namedtuple
-import oracledb
-from ebi_eva_common_pyutils.spreadsheet.metadata_xlsx_utils import metadata_xlsx_version
 
-from convert_gvf_to_vcf.metadataJSON import DGVaMetadataRetriever
+from convert_gvf_to_vcf.metadata_retrievers.evametadataJSON import EVAMetadataRetriever
 
 
 class TestDGVaMetadataRetriever(TestCase):
@@ -17,7 +14,7 @@ class TestDGVaMetadataRetriever(TestCase):
 
     def test_init(self):
         # testing the __init__ function has loaded the config correctly
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         assert metadata_client._host == "mockhost"
         assert metadata_client._port == "1111"
         assert metadata_client._username == "mockuser"
@@ -33,7 +30,7 @@ class TestDGVaMetadataRetriever(TestCase):
     def test_connection_new(self, mock_connection):
         mock_connection_object = Mock()
         mock_connection.return_value = mock_connection_object
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client.connection
         self.assertEqual(result, mock_connection_object)
         mock_connection.assert_called_once()
@@ -46,7 +43,7 @@ class TestDGVaMetadataRetriever(TestCase):
         # expected behaviour: if healthy, keep the connection
         mock_connection_object_existing = Mock()
         mock_connection.return_value = mock_connection_object_existing
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         metadata_client._connection = mock_connection_object_existing
         result = metadata_client.connection
         self.assertEqual(result, mock_connection_object_existing)
@@ -63,7 +60,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_new_connection = Mock()
         mock_connection.return_value = mock_new_connection
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         # set connection to unhealthy
         metadata_client._connection = mock_unhealthy
         result = metadata_client.connection
@@ -82,7 +79,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_healthy.is_healthy.return_value = True
         mock_connection.side_effect = [Exception, Exception, mock_healthy]
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         # checking attribute is set to max 3 retries
         self.assertEqual(metadata_client._max_retries, 3)
         result = metadata_client.connection
@@ -97,7 +94,7 @@ class TestDGVaMetadataRetriever(TestCase):
     def test_connection_max_retry_third_attempt_fail(self, mock_connection):
         # expected behaviour: it will allow 3 retries. it will not allow the fourth attempt
         mock_connection.side_effect = [Exception("attempt1"), Exception("attempt2"), Exception("attempt3"), Exception("attempt4")]
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         with self.assertRaises(Exception) as unsuccessful:
             result = metadata_client.connection
         # this should only allow 3 attempts
@@ -113,7 +110,7 @@ class TestDGVaMetadataRetriever(TestCase):
         expected_data = [('row1',), ('row2',)]
         mock_cursor.fetchall.return_value = expected_data
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client.load_from_db("SQL_QUERY")
         self.assertEqual(result, expected_data)
 
@@ -153,7 +150,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_fetch_all_centres.return_value = mock_data_all_centres
         mock_fetch_all_addresses.return_value = mock_data_all_addresses
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_submitter_details("STUDY123")
 
         # default value only, the None and empty string should only have empty strings for required fields
@@ -178,7 +175,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_fetch_all_addresses.assert_called_once_with("STUDY123")
 
     def test__get_project_pre_registered(self):
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_project_pre_registered("PRJEB123456789")
         expected = {'projectAccession': 'PRJEB123456789'}
         self.assertEqual(expected, result)
@@ -205,7 +202,7 @@ class TestDGVaMetadataRetriever(TestCase):
         centre = mock_centre.return_value[0]
         publications = ['PubMed:12345']
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_project_new("STUDY123")
         expected = {"title": mock_title.return_value,
                     "description": mock_description.return_value,
@@ -229,7 +226,7 @@ class TestDGVaMetadataRetriever(TestCase):
         centre = mock_centre.return_value[0]
         # publications = ['PubMed:12345']
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_project_new("STUDY123")
         expected = {"title": mock_title.return_value,
                     "description": mock_description.return_value,
@@ -278,7 +275,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_software.return_value = ["software1", "software2"]
         mock_pipeline_descriptions.return_value = "my pipeline description"
         mock_run_accessions.return_value = ["ERR123456", "SRR123456"]
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_analysis("estd123", self.vcf_output, mock_reference_fasta, mock_assembly_report)
         expected_result = [{'analysisTitle': 'estd123',
                             'analysisAlias': 'MYanalysisALIAS',
@@ -308,7 +305,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_software.return_value = []
         mock_pipeline_descriptions.return_value = ""
         mock_run_accessions.return_value = []
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_analysis("estd123", self.vcf_output, mock_reference_fasta, mock_assembly_report)
 
         expected_result = [{'analysisTitle': 'estd123',
@@ -324,7 +321,7 @@ class TestDGVaMetadataRetriever(TestCase):
     @patch("convert_gvf_to_vcf.metadataJSON.DGVaMetadataRetriever._fetch_sample_analysis_alias_list")
     def test__get_sample_pre_registered(self, mock_alias_list):
         mock_alias_list.return_value = ["alias"]
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_sample_pre_registered("STUDY123", "SAMNA123456", "favourite_sample")
         expected = {'analysisAlias': ['alias'], 'sampleInVCF': 'favourite_sample', 'bioSampleAccession': 'SAMNA123456'}
         self.assertEqual(result, expected)
@@ -338,7 +335,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_tax_id.return_value = 9606
         mock_scientific_name.return_value = "homo sapiens"
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_sample_new("STUDY123", "favourite_sample")
         expected = {'analysisAlias': ['alias'], 'sampleInVCF': 'favourite_sample', 'bioSampleObject': {'sample_title': 'favourite_sample', 'scientific_name': 'homo sapiens', 'tax_id': 9606, 'collection date': 'not provided', 'geographic location (country and/or sea)': 'not provided'}}
         self.assertEqual(result, expected)
@@ -350,7 +347,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_ids.return_value = ["1", "2"]
         mock_alias.return_value = "alias_1_2"
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._get_files("STUDY123", self.vcf_output)
         expected = [{'analysisAlias': 'alias_1_2', 'fileName': 'a.vcf', 'fileSize': 4177, 'md5': 'a7843773a57dd39a4c85cb7dba59c2c6'}]
         self.assertEqual(result, expected)
@@ -364,14 +361,14 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_data = [(None,) ] # not pre-registered
         mock_load.return_value = mock_data
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         accession = metadata_client._fetch_project_peer_project("STUDY123")
         self.assertEqual(accession, None)
         # create mock data from load_from_db (PROJECT PRE-REGISTERED)
         mock_data = [('PRJNA28889',)] # pre-registered
 
         mock_load.return_value = mock_data
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         accession = metadata_client._fetch_project_peer_project("STUDY789")
         self.assertEqual(accession, "PRJNA28889")
     # convert_gvf_to_vcf.metadataJSON.DGVaMetadataRetriever.load_from_db                TARGET
@@ -384,7 +381,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_data = [ (None, "NA20344"), (None, "NA20345")  ]# not pre-registered
         mock_load.return_value = mock_data
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._determine_sample_pre_registered("estd123")
 
         SampleStatus = namedtuple('SampleStatus', ['is_sample_preregistered', 'sample_accession', 'sample_id'])
@@ -395,7 +392,7 @@ class TestDGVaMetadataRetriever(TestCase):
         mock_data = [("SAMN12345678","mypreregisteredsample")]
         mock_load.return_value = mock_data
 
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._determine_sample_pre_registered("estd123")
 
         SampleStatus = namedtuple('SampleStatus', ['is_sample_preregistered', 'sample_accession', 'sample_id'])
@@ -403,7 +400,7 @@ class TestDGVaMetadataRetriever(TestCase):
         self.assertEqual(result, expected_result)
 
     def test__determine_evidence_type(self):
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._determine_evidence_type(self.vcf_output)
         expected = "genotype"
         self.assertEqual(result, expected)
@@ -411,20 +408,20 @@ class TestDGVaMetadataRetriever(TestCase):
     def test__determine_analysis_experiment_type(self):
         analysis_types= ["Read depth and paired-end mapping"]
         method_types = ["Sequencing"]
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._determine_analysis_experiment_type(analysis_types, method_types)
         expected = "whole_genome_sequencing"
         self.assertEqual(result, expected)
 
 
     def test__format_project(self):
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client._format_project(None, ["www.link.com", "www.link2.com"], None, "123456")
         expected = ('', ['www.link.com| URL', 'www.link2.com| URL'], '', ['PubMed:123456'])
         self.assertEqual(result, expected)
 
     def test_validate_fetch_result(self):
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         # expect string
         result = metadata_client.validate_fetch_result("name", [("value",)], True)
         expected = "value"
@@ -435,7 +432,7 @@ class TestDGVaMetadataRetriever(TestCase):
         self.assertEqual(result, expected)
 
     def test_validate_date(self):
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         result = metadata_client.validate_date("2012-12-12")
         expected = True
         self.assertEqual(result,expected)
@@ -444,7 +441,7 @@ class TestDGVaMetadataRetriever(TestCase):
         self.assertEqual(result, expected)
 
     def test_validate_project(self):
-        metadata_client = DGVaMetadataRetriever(self.config)
+        metadata_client = EVAMetadataRetriever(self.config)
         description = "a description"
         hold_date = "2012-12-12"
         parent = "PRJEA123456"
